+++ scripts/fiddle-test.sh
@@ -4,15 +4,12 @@
 #  Specification           : N/A_____________________________________________________________________|
 #  Specification Path      : N/A_____________________________________________________________________|
 #  Author                  : brady house_____________________________________________________________|
-#  Create date             : 03/19/2015______________________________________________________________|
-#  Description             : UTILITY USED TO "SAFELY" DELETE A FIDDLE SUB-DIRECTORY__________________|
+#  Create date             : 06/21/2015______________________________________________________________|
+#  Description             : UTILITY USED TO RUN JSTESTDRIVER FOR A GIVEN TYPE AND FIDDLE NAME_______|
 #  Command line Arguments  : $1 = FIDDLE TYPE, $2 = FIDDLE NAME______________________________________|
 # ---------------------------------------------------------------------------------------------------|
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
-# 04/15/2015 - See CHANGELOG @ 201504151810
-# 05/01/2015 - See CHANGELOG @ 201505011810
-# 05/08/2015 - See CHANGELOG @ 201505061810
 # 06/20/2015 - See CHANGELOG @ 201506200420
 # ---------------------------------------------------------------------------------------------------|
 clear
@@ -32,17 +29,10 @@ fiddlePath="../fiddles/${fiddleType}/${fiddleName}"
 
     # Verify type parameter
 	case ${fiddleType} in
-        'extjs' | 'jquery' | 'three' | 'php' | 'dojo' | 'chrome' | 'node' | 'tween' | 'bash' )
-        if [[ -d "${fiddlePath}" ]]
-        then
-            sudo rm -r "${fiddlePath}" || exit 87
-        fi
-        case ${fiddleType} in
-            'extjs' | 'jquery' | 'three' | 'php' | 'dojo' | 'tween' )
-                ./fiddle-index.sh ${fiddleType} || exit 88
+        'jquery' )
+            cd bin
+            ./house-jstestdriver-test.sh "../${fiddlePath}"
             ;;
-        esac
-        ;;
         *)
             exit 86
     esac
@@ -50,7 +40,7 @@ fiddlePath="../fiddles/${fiddleType}/${fiddleName}"
 #catch
 _rc=$?
 case ${_rc} in
-    0)  echo "The \"${fiddleName}\" ${fiddleType} fiddle has been deleted successfully."
+    0)  echo "The \"${fiddleName}\" ${fiddleType} fiddle has been tested successfully."
         ;;
     86) echo ""
         echo "Nope ~ Incorrect number of arguments"
@@ -61,15 +51,7 @@ case ${_rc} in
         echo ""
         echo "[t] - type. Valid types include: "
         echo ""
-        echo -e "\t\"bash\"\t\tBash Fiddle"
-        echo -e "\t\"dojo\"\t\tDojo Fiddle"
-        echo -e "\t\"extjs\"\t\tExt JS Fiddle"
-        echo -e "\t\"php\"\t\tPHP Fiddle"
         echo -e "\t\"jquery\"\tjQuery / Bootstrap Fiddle"
-        echo -e "\t\"three\"\t\tthree.js / WebGl Fiddle"
-        echo -e "\t\"chrome\"\tChrome Extension Fiddle"
-        echo -e "\t\"node\"\t\tnode.js Fiddle"
-        echo -e "\t\"tween\"\t\ttween.js Fiddle"
         echo ""
         echo "[n] - fiddle Name.  For example: \"fiddleParabolaSurface\""
         echo ""