+++ scripts/fiddle.sh
@@ -13,6 +13,7 @@
 # Baseline Ver.
 # 05/01/2015 - See CHANGELOG @ 201505011810
 # 05/03/2015 - See CHANGELOG @ 201505310420
+# 06/21/2015 - See CHANGELOG @ 201506210420
 # ---------------------------------------------------------------------------------------------------|
 clear
 thisFile=$(echo "$0" | sed 's/\.\///g')
@@ -50,6 +51,9 @@ echo "${thisFile}" | awk '{print toupper($0)}'
         'refactor')
             if [ "$#" -eq 4 ]; then ./fiddle-refactor.sh $2 $3 $4; fi
             ;;
+        'test')
+            if [ "$#" -eq 3 ]; then ./fiddle-test.sh $2 $3; fi
+            ;;
         *)  exit 86
             ;;
     esac
@@ -75,6 +79,7 @@ case ${_rc} in
         echo -e "\t\"stop\"\t\tStop the web service process"
         echo -e "\t\"delete\"\tDelete an existing fiddle"
         echo -e "\t\"refactor\"\t Rename (\"refactor\") an existing fiddle"
+        echo -e "\t\"test\"\t\tInvoke JsTestDriver for a given fiddle"
         echo ""
         echo -e "[a1-3]\targuments. The arguments required by the "
         echo -e "\tspecified command. There can be up to 3 arguments."